refactor: move image data retrieval to control side
